Split up CodeSearchResults.jsx into multiple components
